requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
# django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
mysqlclient==2.0.3
python-dotenv==0.19.2